feat:deleted api from store and fully moved files and sorted to api.
