+++ api/index.ts
@@ -1,3 +1,8 @@
+import {
+  usePositionApi,
+  usePositionCategoryApi,
+} from './positions/position-client';
+
 export { useApi } from './use-api';
 export { GetAllUsersClient } from './users/users-client';
 export { GetAllUsersServer } from './users/users-server';
@@ -17,3 +22,18 @@ export { GetCompanyOptions } from './companies/companies-client';
 export { GetProfile } from './profile/profile-client';
 
 export { useAuth } from './auth/auth';
+
+export { useCommentApi } from './comments/comment.api';
+export { useCommentTaskApi } from './comments/comment-task.api';
+export { useCommentInterviewApi } from './comments/comment-interview.api';
+export { useCommentInternshipApi } from './comments/comment-internship.api';
+export { useCommentWorkPrimaryApi } from './comments/comment-work-primary.api';
+export { useCommentWorkSecondaryApi } from './comments/comment-work-secondary.api';
+export { useCommentWorkFinanceApi } from './comments/comment-work-finance.api';
+
+export { useProfileApi } from './profile/profile.api';
+
+export { useCompanyStore } from './companies/company.api';
+
+export { usePositionApi } from './positions/position-client';
+export { usePositionCategoryApi } from './positions/position-client';
