+++ app/comments/[id]/page.tsx
@@ -3,7 +3,7 @@
 import { useParams } from 'next/navigation';
 import { Card, Container, Title } from '@/shared/ui';
 import { useEffect } from 'react';
-import { useCommentApi } from '@/store/api';
+import { useCommentApi } from '@/api/index';
 import {
   uniqueNamesGenerator,
   adjectives,
