--- store/api/index.ts
@@ -1,14 +1,0 @@
-export { useCommentApi } from './comment.api';
-export { useCommentTaskApi } from './comment-task.api';
-export { useCommentInterviewApi } from './comment-interview.api';
-export { useCommentInternshipApi } from './comment-internship.api';
-export { useCommentWorkPrimaryApi } from './comment-work-primary.api';
-export { useCommentWorkSecondaryApi } from './comment-work-secondary.api';
-export { useCommentWorkFinanceApi } from './comment-work-finance.api';
-
-export { useProfileApi } from './profile.api';
-
-export { useCompanyStore } from './company.api';
-
-export { usePositionApi } from './position.api';
-export { usePositionCategoryApi } from './position-categories.api';
